Rec: Made Adapter classes file-scoped
